requirements.txt
attrs==23.2.0
certifi==2023.11.17
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
